✨ feat(app): start coding user follow feature
- added all classes need for user follow feature

+++ src/main/java/com/johnmartin/social/dto/response/UserResponse.java
@@ -6,6 +6,6 @@ public record UserResponse(String id,
                            String email,
                            String profileImageUrl,
                            String bio,
-                           int followersNo,
-                           int followingNo) {
+                           long followersNo,
+                           long followingNo) {
 }
